+++ src/com/kaylerrenslow/armaplugin/lang/sqf/psi/SQFStatement.java
@@ -1,9 +1,10 @@
 package com.kaylerrenslow.armaplugin.lang.sqf.psi;
 
+import com.intellij.psi.PsiComment;
 import com.intellij.psi.PsiElement;
 import com.kaylerrenslow.armaplugin.lang.PsiUtil;
 import com.kaylerrenslow.armaplugin.lang.sqf.SQFVariableName;
-import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,12 +20,11 @@ public interface SQFStatement extends PsiElement {
 	 * Gets all private vars declared private in this statement. Note that private variables may be declared private in the
 	 * statement, but not in the same scope as the statement!
 	 * <p>
-	 * This method will return null if the statement is incapable of making variables private. This method will return an empty list
-	 * if the statement can create private variables but none were made private.
+	 * This method will return an empty list if the statement can create private variables but none were made private.
 	 *
-	 * @return list of private vars, or null if can't create private vars
+	 * @return list of private vars
 	 */
-	@Nullable
+	@NotNull
 	default List<SQFPrivateVar> getDeclaredPrivateVars() {
 		SQFScope containingScope = SQFScope.getContainingScope(this);
 		List<SQFPrivateVar> vars = new ArrayList<>();
@@ -37,21 +37,38 @@ public interface SQFStatement extends PsiElement {
 				List<SQFPrivateVar> blockPrivateVars = codeBlockScope.getPrivateVarInstances(false);
 
 				for (SQFPrivateVar blockPrivateVar : blockPrivateVars) {
-					boolean alreadyPrivate = false;
+					boolean varEqual = false;
 					for (SQFPrivateVar statementPrivateVar : vars) {
 						if (blockPrivateVar.getVariableNameObj().nameEquals(statementPrivateVar.getVariableNameObj())
 								&& blockPrivateVar.getMaxScope() == statementPrivateVar.getMaxScope()) {
 							//we need to check scopes as well because a variable may be declared private in a scope
 							//but that may not be it's max scope
-							alreadyPrivate = true;
+							varEqual = true;
 							break;
 						}
 					}
-					if (!alreadyPrivate) {
+					if (!varEqual) {
 						vars.add(blockPrivateVar);
 					}
 				}
 			}
+			return null;
+		};
+
+		/*
+		* Traverse SQFCodeBlock expressions, but not their children
+		* because each call to checkCodeBlockExpr will check the code block's children
+		*/
+		Function<SQFExpression, Void> findAllCodeExpr = sqfExpression -> {
+			PsiUtil.traverseInLayers(sqfExpression.getNode(), astNode -> {
+				PsiElement nodeAsElement = astNode.getPsi();
+				if (nodeAsElement instanceof SQFCodeBlockExpression) {
+					SQFCodeBlockExpression codeBlockExpression = ((SQFCodeBlockExpression) nodeAsElement);
+					checkCodeBlockExp.apply(codeBlockExpression);
+					return true; //don't add children
+				}
+				return false;
+			});
 			return null;
 		};
 
@@ -62,28 +79,12 @@ public interface SQFStatement extends PsiElement {
 			}
 			if (assignment.getExpr() != null) {
 				SQFExpression assignmentExpr = assignment.getExpr();
-				PsiUtil.traverseBreadthFirstSearch(assignmentExpr.getNode(), astNode -> {
-					PsiElement nodeAsPsi = astNode.getPsi();
-					if (nodeAsPsi instanceof SQFCodeBlockExpression) {
-						SQFCodeBlockExpression codeBlockExpression = ((SQFCodeBlockExpression) nodeAsPsi);
-						checkCodeBlockExp.apply(codeBlockExpression);
-						return false; //nothing left to do
-					}
-					if (nodeAsPsi instanceof SQFScope) {
-						//do not traverse further because it was already traversed in SQFCodeBlockExpression
-						return false;
-					}
-					return false;
-				});
+				findAllCodeExpr.apply(assignmentExpr);
 			}
 			return vars;
 		}
 		if (this instanceof SQFExpressionStatement) {
 			SQFExpression expr = ((SQFExpressionStatement) this).getExpr().withoutParenthesis();
-			if (expr instanceof SQFCodeBlockExpression) {
-				checkCodeBlockExp.apply((SQFCodeBlockExpression) expr);
-				return vars;
-			}
 			if (expr instanceof SQFCommandExpression) {
 				SQFCommandExpression cmdExpr = (SQFCommandExpression) expr;
 				switch (cmdExpr.getSQFCommand().getCommandName().toLowerCase()) {
@@ -137,10 +138,53 @@ public interface SQFStatement extends PsiElement {
 					}
 				}
 				return vars;
+			} else {
+				findAllCodeExpr.apply(expr);
 			}
 		}
 
 		//todo we need to check inside code blocks, spawn, and control structures (https://community.bistudio.com/wiki/Variables#Scope)
-		return null;
+		return vars;
+	}
+
+	default boolean isIfStatement() {
+		//todo
+		return false;
+	}
+
+	default boolean isSpawnStatement() {
+		//todo
+		return false;
+	}
+
+	default boolean isSwitchStatement() {
+		//todo
+		return false;
+	}
+
+	default boolean isControlStructure() {
+		//todo
+		return false;
+	}
+
+	/**
+	 * Used for debugging. Will return the statement the given element is contained in.
+	 * If the element is a PsiComment, &lt;PsiComment&gt; will be returned. Otherwise, the element's ancestor statement
+	 * text will be returned with all newlines replaced with spaces.
+	 *
+	 * @return the text, or &lt;PsiComment&gt; if element is a PsiComment
+	 */
+	@NotNull
+	static String getStatementTextForElement(@NotNull PsiElement element) {
+		if (element instanceof PsiComment) {
+			return "<PsiComment>";
+		}
+		if (element.getContainingFile() == null || !(element.getContainingFile() instanceof SQFFile)) {
+			throw new IllegalArgumentException("element isn't in an SQFFile");
+		}
+		while (!(element instanceof SQFStatement)) {
+			element = element.getParent();
+		}
+		return element.getText().replaceAll("\n", " ");
 	}
 }